Search unit tests, PR review changes
