+++ packages/content/rollup.config.js
@@ -1,6 +1,3 @@
 const rollupUtils = require('../../util/rollup-utils');
-const commonJS = require('rollup-plugin-commonjs');
 
-const config = rollupUtils.config('content')
-
-module.exports = config;
+module.exports = rollupUtils.config('content');
