+++ packages/models/src/descriptor.ts
@@ -1,4 +1,5 @@
 export interface Descriptor {
-  page: any;
+  page?: any;
+  component?: any;
   [prop: string]: any;
 }
