+++ packages/models/src/item.ts
@@ -1,8 +1,10 @@
+import { Descriptor } from "./descriptor";
+
 export interface Item {
   name: string;
   url: string;
   descriptorUrl: string;
-  desriptorDom: Object;
+  descriptorDom: Descriptor | null;
   folder: boolean;
   [prop: string]: any;
 }
