+++ packages/redux/src/utils.ts
@@ -30,7 +30,7 @@ export function createReduxStore(config: {
       : combineEpics(...allEpics));
 
   const enhancers = config.reduxDevTools
-    ? ((typeof(window) != "undefined" && window['__REDUX_DEVTOOLS_EXTENSION_COMPOSE__']) || compose)
+    ? ((typeof window !== "undefined" && window['__REDUX_DEVTOOLS_EXTENSION_COMPOSE__']) || compose)
     : compose;
 
   // if config has namespaceCrafterState set to true, combines crafter reducers into namespace, plus config reducers
@@ -104,7 +104,9 @@ export function flattenEntries(item: Item, childrenProperty:string = 'children')
   
   childIds[itemUrl] = [];
   
-  children = item[childrenProperty] ? [ ...item[childrenProperty] ] : [];
+  children = item[childrenProperty] 
+    ? [ ...item[childrenProperty] ] 
+    : [];
 
   //If item has children
   if(children && children.length > 0){
