+++ packages/search/rollup.config.js
@@ -1,3 +1,9 @@
 const rollupUtils = require('../../util/rollup-utils');
+const commonJS = require('rollup-plugin-commonjs');
+const config = rollupUtils.config('search');
 
-module.exports = rollupUtils.config('search');
+config.plugins.push(commonJS({
+    namedExports: { 'uuid': ['uuid'] }
+}));
+
+module.exports = config;
